1.12.2-2.1: changed how to use the "RemovalByKind" CT Command. now it is needed to "get" the wanted "RemovalByKind". use: "val removalByKind = mods.unidict.removalByKind.get("removalByKindName")"" it will return the wanted "RemovalByKind" instance, as long you passed it a valid value. currently the valid values are ("crafting", "furnace") (probably I will have to create a "wiki" to explain better these things XD)
added another "RemovalByKind", the "furnace"
it works exactly as the "crafting" variant, but for the furnace.

after getting the "RemovalByKind" instance, in order to use, use the "remove" method, it uses the same arguments as before.

isn't required to store the "instance" in a val, but if you want to use it several times, it is better.

Happy new Year!